+++ main.go
@@ -4,12 +4,13 @@ package main
 import (
 	"github.com/zrepl/zrepl/cli"
 	"github.com/zrepl/zrepl/client"
+	"github.com/zrepl/zrepl/client/status"
 	"github.com/zrepl/zrepl/daemon"
 )
 
 func init() {
 	cli.AddSubcommand(daemon.DaemonCmd)
-	cli.AddSubcommand(client.StatusCmd)
+	cli.AddSubcommand(status.Subcommand)
 	cli.AddSubcommand(client.SignalCmd)
 	cli.AddSubcommand(client.StdinserverCmd)
 	cli.AddSubcommand(client.ConfigcheckCmd)